Move dependency to pandas into separate modules
