+++ tests/OTAnalytics/plugin_parser/test_otvision_parser.py
@@ -40,10 +40,6 @@ from OTAnalytics.domain.track import (
     TrackRepository,
 )
 from OTAnalytics.domain.video import Video
-from OTAnalytics.plugin_datastore.track_store import (
-    PandasByMaxConfidence,
-    PandasTrackDataset,
-)
 from OTAnalytics.plugin_intersect.simple_intersect import (
     SimpleIntersectBySplittingTrackLine,
 )
@@ -60,14 +56,12 @@ from OTAnalytics.plugin_parser.otvision_parser import (
     CachedVideo,
     CachedVideoParser,
     DetectionFixer,
-    DetectionParser,
     InvalidSectionData,
     OtConfigParser,
     OtEventListParser,
     OtFlowParser,
     OttrkFormatFixer,
     OttrkParser,
-    PandasDetectionParser,
     PythonDetectionParser,
     TrackLengthLimit,
     Version,
@@ -78,7 +72,7 @@ from OTAnalytics.plugin_parser.otvision_parser import (
     _write_bz2,
     _write_json,
 )
-from tests.conftest import TrackBuilder, assert_equal_track_properties
+from tests.conftest import TrackBuilder
 
 
 @pytest.fixture
@@ -244,9 +238,11 @@ class TestOttrkParser:
     def ottrk_parser(
         self, mocked_track_repository: Mock, mocked_track_file_repository: Mock
     ) -> OttrkParser:
-        calculator = PandasByMaxConfidence()
-        detection_parser = PandasDetectionParser(
-            calculator, track_length_limit=DEFAULT_TRACK_LENGTH_LIMIT
+        calculator = ByMaxConfidence()
+        detection_parser = PythonDetectionParser(
+            calculator,
+            mocked_track_repository,
+            track_length_limit=DEFAULT_TRACK_LENGTH_LIMIT,
         )
         return OttrkParser(detection_parser)
 
@@ -269,7 +265,7 @@ class TestOttrkParser:
 
         expected_track = track_builder_setup_with_sample_data.build_track()
 
-        assert result_tracks == PandasTrackDataset.from_list([expected_track])
+        assert result_tracks == PythonTrackDataset.from_list([expected_track])
         ottrk_file.unlink()
 
     def test_parse_bz2(self, example_json_bz2: tuple[Path, dict]) -> None:
@@ -417,67 +413,6 @@ class TestPythonDetectionParser:
         assert d1.video_name == d2.video_name
         assert d1.interpolated_detection == d2.interpolated_detection
         assert d1.track_id == d2.track_id
-
-
-class TestPandasDetectionParser:
-    @pytest.fixture
-    def parser(self) -> DetectionParser:
-        return PandasDetectionParser(
-            PandasByMaxConfidence(),
-            track_length_limit=DEFAULT_TRACK_LENGTH_LIMIT,
-        )
-
-    def test_parse_tracks(
-        self,
-        track_builder_setup_with_sample_data: TrackBuilder,
-        parser: DetectionParser,
-    ) -> None:
-        detections: list[
-            dict
-        ] = track_builder_setup_with_sample_data.build_serialized_detections()
-
-        metadata_video = track_builder_setup_with_sample_data.get_metadata()[
-            ottrk_dataformat.VIDEO
-        ]
-        result_sorted_input = parser.parse_tracks(detections, metadata_video).as_list()
-        unsorted_detections = [detections[-1], detections[0]] + detections[1:-1]
-        result_unsorted_input = parser.parse_tracks(
-            unsorted_detections, metadata_video
-        ).as_list()
-
-        expected_sorted = PandasTrackDataset.from_list(
-            [track_builder_setup_with_sample_data.build_track()]
-        ).as_list()
-
-        for sorted, expected in zip(result_sorted_input, expected_sorted):
-            assert_equal_track_properties(sorted, expected)
-        for unsorted, expected in zip(result_unsorted_input, expected_sorted):
-            assert_equal_track_properties(unsorted, expected)
-
-    @pytest.mark.parametrize(
-        "track_length_limit",
-        [
-            TrackLengthLimit(20, 12000),
-            TrackLengthLimit(0, 4),
-        ],
-    )
-    def test_parse_tracks_consider_minimum_length(
-        self,
-        mocked_track_repository: Mock,
-        track_builder_setup_with_sample_data: TrackBuilder,
-        track_length_limit: TrackLengthLimit,
-    ) -> None:
-        parser = PandasDetectionParser(PandasByMaxConfidence(), track_length_limit)
-        detections: list[
-            dict
-        ] = track_builder_setup_with_sample_data.build_serialized_detections()
-
-        metadata_video = track_builder_setup_with_sample_data.get_metadata()[
-            ottrk_dataformat.VIDEO
-        ]
-        result_sorted_input = parser.parse_tracks(detections, metadata_video).as_list()
-
-        assert len(result_sorted_input) == 0
 
 
 class TestOtFlowParser:
